+++ CHANGELOG.md
@@ -8,7 +8,10 @@
 
 - `SidebarMenuLink` now supports `leftIcon`, `rightIcon`, etc, same as buttons.
 - `SidebarMenuLink` now has `indent` prop and can be used in `SidebarMenuCollapsible`.
+#### `StandardTable`
 
+- Changed name from `headerRowZ` to `headerRowZIndex` option in `StandardTable`.
+- Fix default values for when not setting a `z-index` in configuration (default changed from 1000 to undefined/auto)
 ### Breaking changes
 
 - Removed `SidebarMenuCollapsibleLink`, just use `SidebarMenuLink` with `indent` prop.